Switch to asyncio
Certain codepaths are still run synchronously (such as processing courses).
This patch converts all code paths to asyncio+aiohttp. This improves
throughput but a huge factor.

Note that that all possible options have not yet been tested,
espescially not with a large number of courses. There may be subtle bugs
when operating on a large number of courses.

This patch also adds type annotations to functions.